requirements.txt

# Copyright (c) 2018-2021 Qualcomm Technologies, Inc.

# All rights reserved.

# Redistribution and use in source and binary forms, with or without modification, are permitted (subject to the
# limitations in the disclaimer below) provided that the following conditions are met:

# * Redistributions of source code must retain the above copyright notice, this list of conditions and the following
# disclaimer.
# * Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following
# disclaimer in the documentation and/or other materials provided with the distribution.
# * Neither the name of Qualcomm Technologies, Inc. nor the names of its contributors may be used to endorse or promote
# products derived from this software without specific prior written permission.
# * The origin of this software must not be misrepresented; you must not claim that you wrote the original software.
# If you use this software in a product, an acknowledgment is required by displaying the trademark/log as per the details
# provided here: https://www.qualcomm.com/documents/dirbs-logo-and-brand-guidelines
# * Altered source versions must be plainly marked as such, and must not be misrepresented as being the original software.
# * This notice may not be removed or altered from any source distribution.

# NO EXPRESS OR IMPLIED LICENSES TO ANY PARTY'S PATENT RIGHTS ARE GRANTED BY THIS LICENSE. THIS SOFTWARE IS PROVIDED BY
# THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
# COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
# DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR
# BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
# POSSIBILITY OF SUCH DAMAGE.


flask==1.1.1
requests==2.22.0
elasticsearch==7.0.5
marshmallow==3.2.1
flask_cors==3.0.8
PyYAML==5.1.2
pandas==0.25.3
flask_restful==0.3.7
python-magic==0.4.15
Babel==2.7.0
flask-babel==0.12.2
speaklater==1.3
psycopg2-binary==2.8.3
apispec==3.0.0
flask_apispec==0.8.3
libmagic==1.0
sqlalchemy==1.3.10
flask_script==2.0.6
flask_migrate==2.5.2
psycopg2==2.8.3
werkzeug==0.16.0
click==7.0